+++ CHANGES.md
@@ -1,3 +1,8 @@
+# 0.3
+
+- rename `Arith.mul_quo` to `mul_equo` and fix overflow
+- add `Arith.mul_ediv`
+
 # 0.2
 
 - add `Primes.prime_seq`